Refactor SphericalParticle class to support template dimensions and streamline Mass class methods
